+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix interaction between `--escape-unicode` and `--protect-uppercase` options
 - Protect uppercase option now also protects unicode uppercase
 - Fix using the `--fa` / `--align-values` option leading to empty output